backend/requirements.txt
celery~=5.4.0
django-cors-headers~=4.4.0
django-environ~=0.11.2
Django~=5.0.7
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2
factory-boy~=3.3.0
gunicorn~=22.0.0
pillow~=10.4.0
psycopg2~=2.9.9
